cambios

+++ Desarrollador.java
@@ -2,7 +2,7 @@ public class Desarrollador extends Empleado{
     private double sueldoPorDia;
     private int diasTrabajados;
     private int lenguajes;
-    public Desarrollador(double sueldoPorDia, int diasTrabajados, int lenguajes, int edad, String nombre, int numEmpleado,String correo){
+    public Desarrollador(double sueldoPorDia, int diasTrabajados, int lenguajes, int edad, String nombre, long numEmpleado,String correo){
         super(edad, nombre, numEmpleado,correo);
         this.sueldoPorDia=sueldoPorDia;
         this.diasTrabajados=diasTrabajados;
