+++ Publicidad.java
@@ -2,7 +2,7 @@ public class Publicidad extends Empleado{
     private double sueldoPorDia;
     private int diasTrabajados;
     private int años;
-    public Publicidad(double sueldoPorDia, int diasTrabajados,int añosXP, int edad, String nombre, int numEmpleado,String correo){
+    public Publicidad(double sueldoPorDia, int diasTrabajados,int añosXP, int edad, String nombre, long numEmpleado,String correo){
         super(edad, nombre, numEmpleado,correo);
         this.sueldoPorDia=sueldoPorDia;
         this.diasTrabajados=diasTrabajados;
